pyproject.toml
[project]
name = "mcp-data-assistant"
version = "0.1.0"
description = "Local AI Data Assistant using MCP, FastAPI, and Streamlit"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    # FastAPI & Server
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    
    # MCP (Model Context Protocol)
    "fastmcp>=0.1.0",
    "mcp>=1.0.0",
    
    # AI Client (LM Studio via OpenAI-compatible API)
    "openai>=1.12.0",
    
    # Database
    "pyodbc>=5.1.0",
    
    # File Processing
    "pandas>=2.2.0",
    "openpyxl>=3.1.2",
    
    # Environment & Config
    "python-dotenv>=1.0.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    
    # Logging
    "loguru>=0.7.0",
    
    # HTTP Client
    "httpx>=0.27.0",
    
    # Frontend (Streamlit)
    "streamlit>=1.31.0",
    "streamlit-extras>=0.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "ruff>=0.2.0",
]

[tool.uv]
# UV configuration
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "ruff>=0.2.0",
]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"